+++ gwcelery/tasks/lalinference.py
@@ -10,6 +10,9 @@ import subprocess
 import tempfile
 
 from celery import group
+from glue.lal import Cache
+from gwdatafind import find_urls
+from gwpy.timeseries import StateVector
 import lal
 import lalsimulation
 
@@ -43,6 +46,13 @@ executables = {'datafind': 'gw_data_find',
                'pos_to_sim_inspiral': 'cbcBayesPosToSimInspiral'}
 
 flow = 20.0
+
+# number of samples between end time of the data used for psd estimation and
+# start time of data for PE
+padding = 16
+
+# default number of realizations used for PSD estimation
+default_num_of_realizations = 32
 
 
 def _webdir(graceid):
@@ -95,7 +105,136 @@ def _srate(singleinspiraltable):
            ) * 2
 
 
-def _return_ini(event):
+def _start_end(trigtime, seglen, num_of_realizations):
+    """Return gps start time and end time
+
+    Parameters
+    ----------
+    trigtime : float
+        The time of target trigger
+    seglen : int
+        The length of the segment used to calculate overlap between data and
+        waveform
+    num_of_realizations : int
+        The number of noise realizations used for PSD estimation
+
+    Return
+    ------
+    start : float
+        GPS start time of the whole data used for Parameter Estimation
+    end : float
+        GPS end time of the whole data used for Parameter Estimation
+    """
+    end = trigtime + 2
+    return end - seglen - padding - num_of_realizations * seglen, end
+
+
+def _start_end_of_science_segment_for_one_ifo(start, end, ifo, frametype):
+    """Return gps start time and end time of correctly calibrated and
+    observing-intent data on the time when interferometers are locked. First
+    this function searches for available data. If parts of the data are
+    unavailabile, this function throws away data except for the last
+    continuously available data. Next, this function cheks the Bit 0 (HOFT_OK),
+    1 (OBSERVATION_INTENT) and 2 (OBSERVATION_READY) in GDS-CALIB_STATE_VECTOR
+    and extracts the last continous data which have 1 in all the three bits.
+    Finally this function returns gps start time and end time of the resultant
+    data. This function looks for data only for one specified ifo.
+
+    Parameters
+    ----------
+    start : float
+        GPS start time of the time window over which data is searched for.
+    end : float
+        GPS end time of the time window over which data is searched for.
+    ifo : str
+    frametype : str
+
+    Return
+    ------
+    start : float
+        GPS start time of found science segment
+    end : float
+        GPS end time of found science segment
+    """
+    # get gps start and end time of available data
+    datacache = Cache.from_urls(find_urls(ifo[0], frametype, start, end))
+    available_segment = datacache.to_segmentlistdict()[ifo[0]][-1]
+    start, end = available_segment[0], available_segment[1]
+
+    # check whether data is calibrated correctly, observing-intent and taken
+    # while the inferferometers are locked
+    flag = StateVector.read(
+               datacache, app.conf['state_vector_channel_names'][ifo],
+               start=start, end=end,
+               bits=["HOFT_OK", "OBSERVATION_INTENT", "OBSERVATION_READY"]
+           ).to_dqflags()
+    pe_segment = (flag['HOFT_OK'].active - ~flag['OBSERVATION_INTENT'].active -
+                  ~flag['OBSERVATION_READY'].active)[-1]
+    return pe_segment[0], pe_segment[1]
+
+
+def _start_end_of_science_segment(trigtime, seglen, ifos, frametype_dict):
+    """Calculate gps start time and end time of ready-for-PE data with
+    _start_end_of_science_segment for each ifo and return maximum of start time
+    and minimum of end time"""
+    start, end = _start_end(trigtime, seglen, default_num_of_realizations)
+    for ifo in ifos:
+        start, end = _start_end_of_science_segment_for_one_ifo(
+                         start, end, ifo, frametype_dict[ifo]
+                     )
+    return start, end
+
+
+def _psdstart_psdlength(start, trigtime, seglen):
+    """Return gps start time and length of data for PSD estimation as a list.
+    Note that psdlength has to be seglen multiplied by an integer"""
+    psdlength = \
+        math.floor(trigtime + 2 - seglen - padding - start) // seglen * seglen
+    return [trigtime + 2 - seglen - padding - psdlength, psdlength]
+
+
+def _find_appropriate_frametype_psdstart_psdlength(trigtime, seglen,
+                                                   ifos, superevent_id=None):
+    """Return appropriate frametype, psdstart and psdlength. This function sets
+    long enough psdlength first and shorten psdlength depending on whether data
+    is available, correctly calibrated, observing-intent and taken while the
+    interferometers are locked. This function first searches for low-latency
+    frame data and, if they are not available, searches for high-latency frame
+    data. If enough data is not found finally, raise exception and report
+    failure to GraceDB"""
+    # First search for low-latecy frame data
+    frametype_dict = app.conf['low_latency_frame_types']
+    start, end = _start_end_of_science_segment(
+                     trigtime, seglen, ifos, frametype_dict
+                 )
+    start_threshold, end_threshold = _start_end(trigtime, seglen, 1)
+    # Here we take into account the case where very recent data is not
+    # available at this time but it will be available at the actual start time
+    # of Parameter Estimation
+    if start <= start_threshold:
+        return frametype_dict, _psdstart_psdlength(start, trigtime, seglen)
+
+    # If part of low-latency data has already vanished, search for high-latency
+    # frame data
+    frametype_dict = app.conf['high_latency_frame_types']
+    start, end = _start_end_of_science_segment(
+                     trigtime, seglen, ifos, frametype_dict
+                 )
+    if start <= start_threshold and end >= end_threshold:
+        return frametype_dict, _psdstart_psdlength(start, trigtime, seglen)
+    else:
+        if superevent_id is not None:
+            gracedb.upload.delay(
+                filecontents=None, filename=None,
+                graceid=superevent_id,
+                message='Available data is not long enough, and ' +
+                        'Parameter Estimation will never start automatically.',
+                tags='pe'
+            )
+        raise
+
+
+def _return_ini(event, superevent_id=None):
     """Determine an appropriate PE settings for the target event and return ini
     file content
     """
@@ -104,22 +243,31 @@ def _return_ini(event):
 
     # Download event's info to determine PE settings
     singleinspiraltable = event['extra_attributes']['SingleInspiral']
+    trigtime = event['gpstime']
 
     # fill out the ini template and return the resultant content
     ifos = _ifos(event)
+    seglen = _seglen(singleinspiraltable)
+    # FIXME: seglen here might not be actual seglen if ROQ is used
+    frametypes, psdstart_psdlength = \
+        _find_appropriate_frametype_psdstart_psdlength(
+            trigtime, seglen, ifos, superevent_id
+        )
     ini_settings = {
         'service_url': gracedb.client._service_url,
-        'types': json.dumps(app.conf['frame_types']),
-        'channels': json.dumps(app.conf['channel_names']),
+        'types': json.dumps(frametypes),
+        'channels': json.dumps(app.conf['strain_channel_names']),
         'webdir': _webdir(event['graceid']),
         'paths': [{'name': name, 'path': find_executable(executable)}
                   for name, executable in executables.items()],
         'q': min([sngl['mass2'] / sngl['mass1']
                   for sngl in singleinspiraltable]),
         'ifos': json.dumps(ifos),
-        'seglen': str(_seglen(singleinspiraltable)),
+        'seglen': str(seglen),
         'flow': json.dumps(_freq_dict(flow, ifos)),
-        'srate': str(_srate(singleinspiraltable))
+        'srate': str(_srate(singleinspiraltable)),
+        'psd-start-time': str(psdstart_psdlength[0]),
+        'psd-length': str(psdstart_psdlength[1])
     }
     return ini_template.render(ini_settings)
 
@@ -135,7 +283,7 @@ def upload_ini(event, superevent_id):
     superevent_id : str
         The GraceDb ID of a target superevent
     """
-    ini_contents = _return_ini(event)
+    ini_contents = _return_ini(event, superevent_id)
     gracedb.upload.delay(
         filecontents=ini_contents, filename='online_pe.ini',
         graceid=superevent_id,